+++ src/components/burger-constructor/burger-constructor.module.css
@@ -19,86 +19,4 @@
     flex-direction: column;
     justify-content: flex-start;
     align-items: flex-start;
-    /*outline: 2px solid #4C4CFF;*/
-}
-
-.components {
-    width: 100%;
-    max-height: calc(100vh - 350px);
-    min-height: 100%;
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    align-items: flex-start;
-    justify-content: center;
-}
-
-.components__inside {
-    width: 100%;
-    max-height: calc(100%);
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    align-items: center;
-    overflow-y: auto;
-    overflow-x: hidden;
-}
-
-.components__element {
-    width: 100%;
-    height: 80px;
-    display: block;
-    cursor: move;
-    transition: padding 0.3s ease-in-out;
-}
-
-.components__element_type_bun {
-    width: 536px;
-    height: 80px;
-    margin-left: 30px;
-    flex-grow: 0;
-}
-
-.components__element_bottom {
-    margin-top: auto;
-}
-
-.components__element_type_drag {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    column-gap: 6px;
-    align-items: center;
-}
-
-.components__result {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: center;
-    height: 64px;
-}
-
-.components__price {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-.components__bnt {
-    height: 100%;
-}
-
-.empty {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    background: #1C1C21;
-    border-radius: 30px;
-    box-shadow: 0 24px 32px rgba(0, 0, 0, 0.04), 0 16px 24px rgba(0, 0, 0, 0.04), 0 4px 8px rgba(0, 0, 0, 0.04), 0 0 1px rgba(0, 0, 0, 0.04);
 }
