+++ src/components/burger-constructor/drag-constructor-element/drag-constructor-element.module.css
@@ -1,0 +1,11 @@
+.container {
+    width: 100%;
+    height: 80px;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    column-gap: 6px;
+    align-items: center;
+    cursor: move;
+    transition: padding 0.3s ease-in-out;
+}
